MP out of dyn step 3: remove xtime update from atm_timestep routine
The existing atm_timestep routine calls atm_srk3, and then has an
update of xtime. We are eventually going to remove the physics and
lbc updates from atm_srk3. That means the update of xtime needs to
be removed so that it is at the end of a time step in the
eventual set of calls inside of atm_do_timestep.

Code is removed from routine atm_timestep. That code (the xtime update)
is placed in a short routine in mpas_atm_time_integration.F. The
routine is now called in mpas_atm_core.F in subroutine atm_do_timestep.

modified:   src/core_atmosphere/dynamics/mpas_atm_time_integration.F
modified:   src/core_atmosphere/mpas_atm_core.F